fix saga streamer first part
